refactor(sim): extract the fishing module behind SimContext
Move startFishing/completeFishing and their three private predicates out of
the sim.ts coordinator into src/sim/professions/fishing.ts, bound late with
the live ctx at the host literal (the runEffects idiom). Move-not-rewrite:
emitted strings, guard order, and the single rng draw per normal catch are
byte-identical, so the parity goldens are untouched. sim.ts sheds 111 lines;
the three sim.test.ts reach-ins repoint to the module entry point.
